QUIC: Split  out QUICCryptoStream
