+++ iocore/net/QUICNetVConnection.cc
@@ -55,12 +55,10 @@ static constexpr std::string_view QUIC_DEBUG_TAG = "quic_net"sv;
   Debug("quic_net", "[%s] " fmt, this->cids().data(), ##__VA_ARGS__); \
   Error("quic_net [%s] " fmt, this->cids().data(), ##__VA_ARGS__)
 
-static constexpr uint32_t IPV4_HEADER_SIZE          = 20;
-static constexpr uint32_t IPV6_HEADER_SIZE          = 40;
-static constexpr uint32_t UDP_HEADER_SIZE           = 8;
-static constexpr uint32_t MAX_PACKET_OVERHEAD       = 62; // Max long header len without length of token field of Initial packet
-static constexpr uint32_t MAX_STREAM_FRAME_OVERHEAD = 24;
-// static constexpr uint32_t MAX_CRYPTO_FRAME_OVERHEAD   = 16;
+static constexpr uint32_t IPV4_HEADER_SIZE            = 20;
+static constexpr uint32_t IPV6_HEADER_SIZE            = 40;
+static constexpr uint32_t UDP_HEADER_SIZE             = 8;
+static constexpr uint32_t MAX_PACKET_OVERHEAD         = 62; // Max long header len without length of token field of Initial packet
 static constexpr uint32_t MINIMUM_INITIAL_PACKET_SIZE = 1200;
 static constexpr ink_hrtime WRITE_READY_INTERVAL      = HRTIME_MSECONDS(2);
 static constexpr uint32_t PACKET_PER_EVENT            = 256;
